refractor: seperate threads into its own script
