+++ w3d_translator/w3d_translator.py
@@ -5,7 +5,7 @@ import os
 
 from unity import UNITY_VERSION, UNITY_PATH
 from validate import validate_input, validate_output
-from xml_to_unity import projectThread
+from threads import projectThread
 
 # TODO: Configure typer character length (100)
 # TODO: Configure color/style for typer.echo
